혼밥 만족도 바로 표시(계산식 포함) + 리뷰 작성 기능 및 리뷰수 자동 증가 + 리뷰 DB등록

+++ src/main/java/com/app/dto/restaurant/ReviewDTO.java
@@ -7,8 +7,8 @@ public class ReviewDTO {
 	private Long id;
     private Long restaurantId;
     private Long userId;
-    private Double rating;
-    private Double soloScore;
+    private Integer rating;
+    private Integer soloScore;
     private String reviewComment;
     private String status;
     private String createdAt;
